alert-ingestion-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
prometheus-client==0.19.0
prometheus-fastapi-instrumentator==6.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
httpx==0.26.0
pydantic==2.5.3
pydantic-settings==2.1.0
